fixed issue where menus will stick out of the right click menu box
